% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean raw NOAA data frame}
\usage{
eq_clean_data(dt)
}
\arguments{
\item{dt}{A raw NOAA data frame}
}
\value{
The function returns a clean data frame with DATE, LATITUDE and LONGITUDE
columns in appropriate format.
}
\description{
This function takes raw NOAA data frame and returns a clean data frame.
The clean data frame have a DATE column created by uniting the year, month and
day columns. As well LATITUDE and LONGITUDE columns are converted to numeric class.
}
\examples{
\dontrun{
eq_clean_data(data.table::fread(file.path("inst/extdata", "signif.txt"), sep = "\\t"))
}

}
