% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean the LOCATION_NAME column in NOAA data frame}
\usage{
eq_location_clean(dt)
}
\arguments{
\item{dt}{A raw NOAA data frame}
}
\value{
The function returns a cleaned up version of the LOCATION_NAME column.
}
\description{
This function cleans the LOCATION_NAME column by stripping out the country name
and converts names to title case.
}
\examples{
\dontrun{
eq_location_clean(data.table::fread(file.path("inst/extdata", "signif.txt"), sep = "\\t"))
}

}
