% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Build interactive chart}
\usage{
eq_map(dt, annot_col = "DATE")
}
\arguments{
\item{dt}{NOAA data frame}

\item{annot_col}{A character vector containing the name of a column which is used for annotation
in the pop-up window}
}
\value{
The function returns an interactive map with annotation in popup window
}
\description{
This function visualizes earthquakes on an interactive map. The function maps the
epicenters (LATITUDE/LONGITUDE) and annotates each point with in pop up window
containing annotation data stored in a column of the data frame.
}
\examples{
\dontrun{

readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")

}

}
