% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Create a geom for plotting a time line of NOAA earthquakes}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
GeomTimeline
}
\description{
Create a ggproto object for plotting a time line of earthquakes.
}
\examples{
\dontrun{

  dt[YEAR > 2000 & COUNTRY \%in\% c("USA", 'CHINA'), ] \%>\%
  ggplot2::ggplot(aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, color = DEATHS)) +
  geom_timeline() +
  theme_timeline()

}

}
\keyword{datasets}
