% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Create a geom for adding a text annotation to a time line plot}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
Create a geom for adding a vertical line to each data point on a time line
chart with a text annotation.
}
\examples{
\dontrun{

  dt[YEAR > 2000 & COUNTRY \%in\% c("USA", 'CHINA'), ] \%>\%
  ggplot2::ggplot(aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, color = DEATHS)) +
  geom_timeline() +
  geom_timeline_label(aes(label = LOCATION_NAME), n_max = 5) +
  theme_timeline()

}

}
\keyword{datasets}
