% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Time line theme}
\usage{
theme_timeline()
}
\description{
Create custom theme for time line chart.
}
\examples{
\dontrun{

  dt[YEAR > 2000 & COUNTRY \%in\% c("USA", 'CHINA'), ] \%>\%
  ggplot2::ggplot(aes(x = DATE, y = COUNTRY, size = EQ_PRIMARY, color = DEATHS)) +
  geom_timeline() +
  theme_timeline()

}

}
